+++ lib/src/clojure_lsp/settings.clj
@@ -70,11 +70,17 @@
     (into ["powershell.exe" "-NoProfile"] classpath)
     classpath))
 
-(defn ^:private default-project-specs []
+(defn ^:private deps-source-aliases [source-aliases]
+  (let [aliases (if source-aliases
+                  (map name source-aliases)
+                  ["dev" "test"])]
+    (str "-A:" (string/join ":" aliases))))
+
+(defn ^:private default-project-specs [source-aliases]
   (->> [{:project-path "project.clj"
          :classpath-cmd ["lein" "classpath"]}
         {:project-path "deps.edn"
-         :classpath-cmd ["clojure" "-A:dev:test" "-Spath"]}
+         :classpath-cmd ["clojure" (deps-source-aliases source-aliases) "-Spath"]}
         {:project-path "build.boot"
          :classpath-cmd ["boot" "show" "--fake-classpath"]}
         {:project-path "shadow-cljs.edn"
@@ -95,14 +101,15 @@
   [old-settings new-settings project-root-uri env]
   (cond-> new-settings
 
-    (setting-changed? old-settings new-settings :project-specs env)
-    (update :project-specs #(or % (default-project-specs)))
-
     (setting-changed? old-settings new-settings :source-aliases env)
     (update :source-aliases #(or % source-paths/default-source-aliases))
 
     (setting-changed? old-settings new-settings :source-paths env)
-    (update :source-paths (partial source-paths/process-source-paths (shared/uri->path project-root-uri) new-settings))))
+    (update :source-paths (partial source-paths/process-source-paths (shared/uri->path project-root-uri) new-settings))
+
+    (setting-changed? old-settings new-settings :project-specs env)
+    (update :project-specs #(or % (default-project-specs (or (:source-aliases new-settings)
+                                                             (:source-aliases old-settings)))))))
 
 (defn ^:private get-refreshed-settings [db]
   (let [{:keys [project-root-uri settings force-settings env]} @db